producer/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
confluent-kafka
